Move entrytype dialog to distinct package
