+++ src/main/java/org/jabref/gui/entrytype/EntryType.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.layout.VBox?>
 <DialogPane xmlns:fx="http://javafx.com/fxml/1" prefWidth="476.0" xmlns="http://javafx.com/javafx/8.0.171"
-            fx:controller="org.jabref.gui.EntryTypeView">
+            fx:controller="org.jabref.gui.entrytype.EntryTypeView">
     <buttonTypes>
         <ButtonType fx:constant="CANCEL"/>
         <ButtonType fx:id="generateButton" buttonData="OK_DONE" text="%Generate"/>
