+++ src/main/java/org/jabref/gui/entrytype/EntryTypeView.java
@@ -1,4 +1,4 @@
-package org.jabref.gui;
+package org.jabref.gui.entrytype;
 
 import java.util.Collection;
 import java.util.List;
@@ -17,6 +17,10 @@ import javafx.scene.control.Tooltip;
 import javafx.scene.layout.FlowPane;
 import javafx.stage.Screen;
 
+import org.jabref.gui.DialogService;
+import org.jabref.gui.Globals;
+import org.jabref.gui.LibraryTab;
+import org.jabref.gui.StateManager;
 import org.jabref.gui.util.BaseDialog;
 import org.jabref.gui.util.ControlHelper;
 import org.jabref.gui.util.IconValidationDecorator;
