+++ src/main/java/org/jabref/gui/entrytype/EntryTypeViewModel.java
@@ -1,4 +1,4 @@
-package org.jabref.gui;
+package org.jabref.gui.entrytype;
 
 import java.util.Optional;
 
@@ -14,6 +14,9 @@ import javafx.collections.FXCollections;
 import javafx.concurrent.Task;
 import javafx.concurrent.Worker;
 
+import org.jabref.gui.DialogService;
+import org.jabref.gui.LibraryTab;
+import org.jabref.gui.StateManager;
 import org.jabref.gui.externalfiles.ImportHandler;
 import org.jabref.gui.importer.NewEntryAction;
 import org.jabref.gui.util.TaskExecutor;
