+++ src/main/java/org/jabref/gui/importer/NewEntryAction.java
@@ -6,12 +6,12 @@ import java.util.Optional;
 import java.util.function.Supplier;
 
 import org.jabref.gui.DialogService;
-import org.jabref.gui.EntryTypeView;
 import org.jabref.gui.LibraryTab;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.Telemetry;
 import org.jabref.gui.actions.ActionHelper;
 import org.jabref.gui.actions.SimpleCommand;
+import org.jabref.gui.entrytype.EntryTypeView;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.types.EntryType;
 import org.jabref.preferences.PreferencesService;
